Remove old, unused files and move everything to the top-level dir
