+++ src/meson.build
@@ -22,10 +22,10 @@ applet_trash_deps = [
 ]
 
 applet_trash_sources = [
-    'ui/trash_icon_button.c',
-    'ui/trash_item.c',
-    'ui/trash_revealer.c',
-    'ui/trash_store.c',
+    'trash_icon_button.c',
+    'trash_item.c',
+    'trash_revealer.c',
+    'trash_store.c',
     'utils.c',
     'applet.c',
     'plugin.c',
