--- src/trash_handler.h
@@ -1,6 +1,0 @@
-#ifndef __BTA_TRASH_HANDLER_H
-#define __BTA_TRASH_HANDLER_H
-
-
-
-#endif
