+++ src/trash_icon_button.c
