+++ src/trash_icon_button.h
