+++ src/trash_revealer.c
