+++ src/trash_revealer.h
